requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-hashes --format requirements.txt -o requirements.txt
altair==6.0.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via httpx
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
beautifulsoup4==4.14.3
    # via yfinance
blinker==1.9.0
    # via streamlit
cachetools==6.2.6
    # via streamlit
certifi==2026.1.4
    # via
    #   curl-cffi
    #   httpcore
    #   httpx
    #   oci
    #   requests
cffi==2.0.0
    # via
    #   cryptography
    #   curl-cffi
charset-normalizer==3.4.4
    # via requests
circuitbreaker==2.1.3
    # via oci
click==8.3.1
    # via streamlit
colorama==0.4.6 ; sys_platform == 'win32'
    # via
    #   click
    #   tqdm
cryptography==45.0.7
    # via
    #   google-auth
    #   oci
    #   pyopenssl
curl-cffi==0.13.0
    # via yfinance
frozendict==2.4.7
    # via yfinance
gitdb==4.0.12
    # via gitpython
gitpython==3.1.46
    # via streamlit
google-ai-generativelanguage==0.6.15
    # via google-generativeai
google-api-core==2.25.2
    # via
    #   google-ai-generativelanguage
    #   google-api-python-client
    #   google-generativeai
google-api-python-client==2.190.0
    # via google-generativeai
google-auth==2.48.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-generativeai
google-auth-httplib2==0.3.0
    # via google-api-python-client
google-generativeai==0.8.6
    # via crypto-monitor
googleapis-common-protos==1.72.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.78.1
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.71.2
    # via google-api-core
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via
    #   httpx
    #   python-telegram-bot
httplib2==0.31.2
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httpx==0.28.1
    # via python-telegram-bot
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
jsonschema==4.26.0
    # via altair
jsonschema-specifications==2025.9.1
    # via jsonschema
markupsafe==3.0.3
    # via jinja2
multitasking==0.0.12
    # via yfinance
narwhals==2.16.0
    # via
    #   altair
    #   plotly
numpy==2.4.2
    # via
    #   pandas
    #   pydeck
    #   streamlit
    #   ta
    #   yfinance
oci==2.167.2
    # via crypto-monitor
packaging==26.0
    # via
    #   altair
    #   plotly
    #   streamlit
pandas==2.3.3
    # via
    #   crypto-monitor
    #   streamlit
    #   ta
    #   yfinance
peewee==4.0.0
    # via yfinance
pillow==12.1.1
    # via streamlit
platformdirs==4.9.2
    # via yfinance
plotly==6.5.2
    # via crypto-monitor
proto-plus==1.27.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==5.29.6
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-generativeai
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
    #   streamlit
    #   yfinance
pyarrow==23.0.1
    # via streamlit
pyasn1==0.6.2
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==3.0 ; implementation_name != 'PyPy'
    # via cffi
pydantic==2.12.5
    # via google-generativeai
pydantic-core==2.41.5
    # via pydantic
pydeck==0.9.1
    # via streamlit
pyopenssl==25.1.0
    # via oci
pyparsing==3.3.2
    # via httplib2
python-dateutil==2.9.0.post0
    # via
    #   oci
    #   pandas
python-telegram-bot==22.6
    # via crypto-monitor
pytz==2025.2
    # via
    #   oci
    #   pandas
    #   yfinance
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via
    #   crypto-monitor
    #   google-api-core
    #   streamlit
    #   yfinance
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
rsa==4.9.1
    # via google-auth
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
soupsieve==2.8.3
    # via beautifulsoup4
streamlit==1.54.0
    # via crypto-monitor
ta==0.11.0
    # via crypto-monitor
tenacity==9.1.4
    # via streamlit
toml==0.10.2
    # via streamlit
tornado==6.5.4
    # via streamlit
tqdm==4.67.3
    # via google-generativeai
typing-extensions==4.15.0
    # via
    #   altair
    #   anyio
    #   beautifulsoup4
    #   google-generativeai
    #   grpcio
    #   pydantic
    #   pydantic-core
    #   pyopenssl
    #   referencing
    #   streamlit
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.3
    # via pandas
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.6.3
    # via requests
watchdog==6.0.0 ; sys_platform != 'darwin'
    # via streamlit
websockets==16.0
    # via yfinance
yfinance==1.2.0
    # via crypto-monitor
